Improvement for [HER-1] Support 'Crawl-Delay' and 'Allow' robots.txt directives * AbstractFrontier.java     change boolean 'respect-crawl-delay' to integer 'respect-crawl-delay-up-to-secs' Fix for [HER-1516] in datamodel.Robotstxt, Integer.parseInt generates NumberFormatException on floating point Crawl-delay strings * RobotsDirectives.java, RobotsTxt.java, RobotsExclusionPolicy.java     parse, store, return crawlDelay as float

+++ trunk/heritrix/src/java/org/archive/crawler/datamodel/RobotsExclusionPolicy.java
@@ -257,7 +257,7 @@ public class RobotsExclusionPolicy implements Serializable {
      * @param userAgent
      * @return int Crawl-Delay value, or -1 if non available
      */
-    public int getCrawlDelay(String userAgent) {
+    public float getCrawlDelay(String userAgent) {
         if (robotstxt==null) {
             return -1;
         }
